# Optimize declarative environment record handling for functions