--- a/Jint.Benchmark/Properties/AssemblyInfo.cs
+++ b/Jint.Benchmark/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following