… Jint/Native/Object/ObjectConstructor.cs
… Jint/Native/String/StringPrototype.cs
… Jint/Options.cs
… Jint/Pooling/ObjectPool.cs
… Jint/Runtime/Debugger/BreakPoint.cs
… Jint/Runtime/Descriptors/PropertyDescriptor.cs
… Jint/Runtime/Environments/DeclarativeEnvironmentRecord.cs
… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/Interop/TypeReference.cs
… Jint/Runtime/MruPropertyCache2.cs
… Jint/Runtime/References/Reference.cs
… Jint/Runtime/StatementInterpreter.cs
… Jint/Runtime/TypeConverter.cs